requirements.txt
qdrant-client
langchain
openai
fastapi
uvicorn
python-dotenv
langchain-community
langchain-openai
python-multipart